Removed nb_1 and nb_sigma1 variables and knots_1 & knots_sigma_1 to scalar.
It seems that when converting data files to json format, we might confuse
a vector of length 1 and a scalar. For example if nb_1 = 1 then knots_1
should be a vector[1] but when running posterior_reference Stan interprets
the type of data for knots_1 and knots_sigma_1 as scalar and error is produced.